pyproject.toml
[tool.poetry]
name = "fashion-image-classifier"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Scrapy = "^2.4.1"
scrapy-splash = "^0.7.2"
python-dotenv = "^0.15.0"
SQLAlchemy = "^1.3.23"
pandas = "^1.2.1"
Keras = "^2.4.3"
tqdm = "^4.56.2"
bayesian-optimization = "^1.2.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
ipython = "^7.20.0"
mypy = "^0.800"
black = "^20.8b1"
pycodestyle = "^2.6.0"
Pillow = "^8.1.0"
pydocstyle = "^5.1.1"
jupyterlab = "^3.0.7"
pylama = "^7.7.1"
pigeon-jupyter = "^0.1.0"
sklearn = "^0.0"
matplotlib = "^3.3.4"
seaborn = "^0.11.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
